Update gradio_app.py
I messed up. BAD.

The changes i implemented initiall, whilst they DID allow me to massivley increase the quality of the output file, for some reason.. it also caused the whole program to run like garbage, even on the default settings, regardless of the options.. which is a bit dumb on my part. I thought the slowdown was purely cos of me running it harder. Very clearly not!

fooled around a bit and made some more changes (probably a ton more than I needed lol) and it does appear to exhibit correct behaviour and is WAAAY faster now. 

Sorry for this total messup!!